Solution to Calculate and Print the Area of a Circle

+++ Calculate and Print the Area of a Circle@@6788d5012db25efec946f943/code.c
@@ -2,7 +2,7 @@
 int main()
 {
     float a;
-    int area;
+    float area;
     scanf("%f",&a);
     area=(3.14)*a*a;
     printf("Area: %.2f",area);
